refactor: WebUIChessApplication 클래스 내 render를 WebController로 이동
